requirements.txt
requests
psycopg2-binary
geocoder==1.38.1
python-dotenv
pydantic-settings
